feat(main): move to internal package
BREAKING CHANGE: This moves out of main and uses errors.
