+++ main.go
@@ -1,0 +1,19 @@
+// gocovmerge takes the results from multiple `go test -coverprofile` runs and
+// merges them into one profile
+package main
+
+import (
+	"flag"
+	"log"
+	"os"
+
+	"github.com/alexfalkowski/gocovmerge/internal/cmd"
+)
+
+func main() {
+	flag.Parse()
+
+	if err := cmd.Run(os.Stdout, flag.Args()); err != nil {
+		log.Fatal(err)
+	}
+}
